requirements.txt
sounddevice
numpy
librosa
soundfile
pydantic
requests
python-dotenv
